kernel: fix 20 code-review findings across net, mm, boot, kernel, drivers, fs, Rust
Critical/major:
- tcp: ACK range check used zero-extending long casts, accepting stale
  ACKs (send-buffer loss) and ACKs above SndNxt (inFlight underflow
  wedging the connection); use 32-bit signed casts
- mm: plain operator new returned nullptr but the compiler assumes it
  cannot (noexcept redeclaration is rejected; clang 14 ignores
  -fcheck-new), so constructors ran on null and OOM checks were dead.
  Plain new now panics on OOM; fallible sites use new (Mm::NoThrow) T,
  which the compiler must null-check (29 sites converted)
- boot: AP trampoline ran 32-bit code on the stale real-mode SS/DS, so
  the first call pushed into the live boot page tables; load gdt32.data
  into SS/DS/ES on protected-mode entry
- kernel: per-CPU TSS with IST1 stack, ltr on BSP+APs, #DF gate on IST1
  so double faults panic with diagnostics instead of triple-faulting
- cpu: SetBspIndex marked CpuArray[0] running instead of CpuArray[index]
- cmd: partitions DMA'd a full hardware sector into a 512-byte stack
  buffer (4K-LBA NVMe overflows); use a device-sized heap buffer
- icxxabi: __cxa_guard_acquire leaked PreemptDisable on the lost-race
  path, permanently pinning the losing task to its CPU
- tsc: kvmclock sampled the pvclock entry with migration possible and
  never checked PVCLOCK_TSC_STABLE_BIT; sample with IRQs off, clamp
  delta underflow, enforce cross-CPU monotonicity via atomic max
- mm: page-allocator MapPage rollbacks skipped the cross-CPU TLB
  shootdown before recycling the VA block and pages
- ext2: mount accepted ext3/ext4 images (same magic) and returned
  garbage reads as success; refuse unknown incompat bits, require
  filetype, bound LogBlockSize before shifting; Unmount frees vnodes by
  tree walk (cache-overflow leak); hard-link/cyclic dirents no longer
  rename/re-parent an already-linked vnode

Minor:
- rust_ffi: kernel_interrupt_register_level reported refused
  registrations as success (driver left deaf); roll back slot, return 0
- rust: arm TCO watchdog only after the kick timer is secured
- virtio_net: clear TxSlotByHead on completion (double-free on
  duplicated used-ring entries)
- virtio_rng: quarantine timed-out requests instead of misattributing
  late completions and leaking descriptors
- virtio_pci: bound the vendor-capability walk (cyclic list hang)
- acpi: map RSDT header straddle-aware, remap full table before the
  checksum walk
- format: accept %ll length modifier (lines with %llu were dropped
  whole); cast u32/bool trace args to ulong at mismatched call sites
- nanofs: re-mark tree-reachable inodes/blocks in the allocation
  bitmaps at mount so a crafted image cannot make AllocInode/
  AllocDataBlock clobber live metadata

Verified: full docker build (clang -Werror), cppcheck clean, QEMU boot
with 2 CPUs -- self-tests pass, both CPUs schedule tasks, DHCP up,
shell prompt reached, no panics.

+++ src/cpp/lib/btree.h
@@ -392,7 +392,7 @@ private:
         bool IsFull()
         {
             bool result = ((2 * T - 1) == KeyCount) ? true : false;
-            Trace(BtreeLL, "node 0x%p full %d keyCount %lu", this, result, KeyCount);
+            Trace(BtreeLL, "node 0x%p full %u keyCount %lu", this, (ulong)result, KeyCount);
             return result;
         }
 
@@ -565,7 +565,7 @@ private:
         void SetLeaf(bool leaf)
         {
             Leaf = leaf;
-            Trace(BtreeLL, "node 0x%p set leaf %d", this, leaf);
+            Trace(BtreeLL, "node 0x%p set leaf %u", this, (ulong)leaf);
         }
 
         void SetKeyCount(size_t keyCount)
